Backend: SubCategory and ChildCategory route protection done

+++ app/Http/Controllers/Backend/ChildCategoryController.php
@@ -13,12 +13,23 @@ use App\Models\ChildCategory;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 use Yajra\DataTables\DataTables;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Blade;
+use Spatie\Permission\Exceptions\UnauthorizedException;
 
 
 class ChildCategoryController extends Controller
 {
     use ImageUploadTraits;
 
+    public $user;
+    public function __construct()
+    {
+        $this->user = Auth::guard('admin')->user();
+        if (!$this->user) {
+            abort(403, 'Unauthorized access');
+        }
+    }
      /**
      * Display a listing of the resource.
      */
@@ -41,48 +52,56 @@ class ChildCategoryController extends Controller
             ->get();
 
         return DataTables::of($childCategories)
-
             ->addIndexColumn()
             ->addColumn('childCategoryImg', function ($childCategory) {
                 return '<a href="'.asset( $childCategory->img ).'" target="__blank">
                     <img src="'.asset( $childCategory->img ).'" width="50px" height="50px">
                 </a>';
             })
-
             ->addColumn('status', function ($childCategory) {
-                if ($childCategory->status == 1) {
-                    return ' <a class="status" id="status" href="javascript:void(0)"
-                        data-id="'.$childCategory->id.'" data-status="'.$childCategory->status.'"> <i
-                            class="fa-solid fa-toggle-on fa-2x"></i>
-                    </a>';
-                } else {
-                    return '<a class="status" id="status" href="javascript:void(0)"
-                        data-id="'.$childCategory->id.'" data-status="'.$childCategory->status.'"> <i
-                            class="fa-solid fa-toggle-off fa-2x" style="color: grey"></i>
-                    </a>';
+                if(auth("admin")->user()->can("status.childcategory"))
+                    if ($childCategory->status == 1) {
+                        return ' <a class="status" id="status" href="javascript:void(0)"
+                            data-id="'.$childCategory->id.'" data-status="'.$childCategory->status.'"> <i
+                                class="fa-solid fa-toggle-on fa-2x"></i>
+                        </a>';
+                    } else {
+                        return '<a class="status" id="status" href="javascript:void(0)"
+                            data-id="'.$childCategory->id.'" data-status="'.$childCategory->status.'"> <i
+                                class="fa-solid fa-toggle-off fa-2x" style="color: grey"></i>
+                        </a>';
+                    }
+                else{
+                    return '<span class="badge bg-info">N/A</span>'; 
                 }
             })
 
             ->addColumn('action', function ($childCategory) {
-                return '
-                <div class="btn-group">
-                    <button type="button" class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">Actions <i class="mdi mdi-chevron-down"></i>
-                    </button>
+                $actionHtml = Blade::render('
+                    <div class="btn-group">
+                        <button type="button" class="btn btn-primary dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">Actions <i class="mdi mdi-chevron-down"></i>
+                        </button>
 
-                    <div class="dropdown-menu dropdownmenu-primary" style="">
-                        <a class="dropdown-item text-info" id="viewButton" href="javascript:void(0)" data-id="'.$childCategory->id.'" data-bs-toggle="modal" data-bs-target="#viewModal">
-                            <i class="fas fa-eye"></i> View
-                        </a>
+                        <div class="dropdown-menu dropdownmenu-primary" style="">
+                            <a class="dropdown-item text-info" id="viewButton" href="javascript:void(0)" data-id="'.$childCategory->id.'" data-bs-toggle="modal" data-bs-target="#viewModal">
+                                <i class="fas fa-eye"></i> View
+                            </a>
 
-                        <a class="dropdown-item text-success" id="editButton" href="javascript:void(0)" data-id="'.$childCategory->id.'" data-bs-toggle="modal" data-bs-target="#editModal">
-                            <i class="fas fa-edit"></i> Edit
-                        </a>
+                            @if(auth("admin")->user()->can("update.childcategory"))
+                                <a class="dropdown-item text-success" id="editButton" href="javascript:void(0)" data-id="'.$childCategory->id.'" data-bs-toggle="modal" data-bs-target="#editModal">
+                                    <i class="fas fa-edit"></i> Edit
+                                </a>
+                            @endif
 
-                        <a class="dropdown-item text-danger" href="javascript:void(0)" data-id="'.$childCategory->id.'" id="deleteBtn">
-                            <i class="fas fa-trash"></i> Delete
-                        </a>
+                            @if(auth("admin")->user()->can("delete.childcategory"))
+                                <a class="dropdown-item text-danger" href="javascript:void(0)" data-id="'.$childCategory->id.'" id="deleteBtn">
+                                    <i class="fas fa-trash"></i> Delete
+                                </a>
+                            @endif
+                        </div>
                     </div>
-                </div>';
+                ', ['childCategory' => $childCategory]);
+                return $actionHtml;
             })
 
             ->rawColumns(['childCategoryImg','status','action'])
@@ -94,6 +113,10 @@ class ChildCategoryController extends Controller
      */
     public function store(CreateChildCategoryRequest $request)
     {   
+        if (!$this->user || !$this->user->can('create.childcategory')) {
+            throw UnauthorizedException::forPermissions(['create.childcategory']);
+        }
+
         DB::beginTransaction();
         try {
 
@@ -128,6 +151,10 @@ class ChildCategoryController extends Controller
      */
     public function changeChildCategoryStatus(Request $request)
     {
+        if (!$this->user || !$this->user->can('status.childcategory')) {
+            throw UnauthorizedException::forPermissions(['status.childcategory']);
+        }
+
         $id = $request->id;
         $Current_status = $request->status;
 
@@ -151,6 +178,10 @@ class ChildCategoryController extends Controller
      */
     public function edit(ChildCategory $childCategory)
     {
+        if (!$this->user || !$this->user->can('update.childcategory')) {
+            throw UnauthorizedException::forPermissions(['update.childcategory']);
+        }
+
         // dd($childCategory);
         return response()->json(['success' => $childCategory]);
     }
@@ -160,6 +191,10 @@ class ChildCategoryController extends Controller
      */
     public function update(UpdateChildCategoryRequest $request, string $id)
     {
+        if (!$this->user || !$this->user->can('update.childcategory')) {
+            throw UnauthorizedException::forPermissions(['update.childcategory']);
+        }
+
         $childCategory = ChildCategory::find($id);
 
         DB::beginTransaction();
@@ -192,6 +227,10 @@ class ChildCategoryController extends Controller
      */
     public function destroy(ChildCategory $childCategory)
     {
+        if (!$this->user || !$this->user->can('delete.childcategory')) {
+            throw UnauthorizedException::forPermissions(['delete.childcategory']);
+        }
+
         if ($childCategory->img) {
             if (file_exists($childCategory->img)) {
                 unlink($childCategory->img);
